Refactor dto package to entities
