+++ app/src/main/java/app/dependency/update/app/controller/MongoRepoController.java
@@ -6,10 +6,10 @@ import app.dependency.update.app.model.MongoDependencies;
 import app.dependency.update.app.model.MongoNpmSkips;
 import app.dependency.update.app.model.MongoPackages;
 import app.dependency.update.app.model.MongoPlugins;
-import app.dependency.update.app.model.dto.Dependencies;
-import app.dependency.update.app.model.dto.NpmSkips;
-import app.dependency.update.app.model.dto.Packages;
-import app.dependency.update.app.model.dto.Plugins;
+import app.dependency.update.app.model.entities.Dependencies;
+import app.dependency.update.app.model.entities.NpmSkips;
+import app.dependency.update.app.model.entities.Packages;
+import app.dependency.update.app.model.entities.Plugins;
 import app.dependency.update.app.service.MongoRepoService;
 import io.swagger.v3.oas.annotations.Operation;
 import java.util.List;
