+++ app/src/main/java/app/dependency/update/app/repository/DependenciesRepository.java
@@ -1,6 +1,6 @@
 package app.dependency.update.app.repository;
 
-import app.dependency.update.app.model.dto.Dependencies;
+import app.dependency.update.app.model.entities.Dependencies;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
