+++ app/src/main/java/app/dependency/update/app/repository/PluginsRepository.java
@@ -1,6 +1,6 @@
 package app.dependency.update.app.repository;
 
-import app.dependency.update.app.model.dto.Plugins;
+import app.dependency.update.app.model.entities.Plugins;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
