+++ app/src/main/java/app/dependency/update/app/runnable/ExecuteGradleUpdate.java
@@ -10,8 +10,8 @@ import app.dependency.update.app.model.GradleDependency;
 import app.dependency.update.app.model.GradlePlugin;
 import app.dependency.update.app.model.Repository;
 import app.dependency.update.app.model.ScriptFile;
-import app.dependency.update.app.model.dto.Dependencies;
-import app.dependency.update.app.model.dto.Plugins;
+import app.dependency.update.app.model.entities.Dependencies;
+import app.dependency.update.app.model.entities.Plugins;
 import app.dependency.update.app.service.MongoRepoService;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
