+++ app/src/main/java/app/dependency/update/app/service/MongoRepoService.java
@@ -6,10 +6,10 @@ import app.dependency.update.app.connector.MavenConnector;
 import app.dependency.update.app.model.MavenDoc;
 import app.dependency.update.app.model.MavenResponse;
 import app.dependency.update.app.model.MavenSearchResponse;
-import app.dependency.update.app.model.dto.Dependencies;
-import app.dependency.update.app.model.dto.NpmSkips;
-import app.dependency.update.app.model.dto.Packages;
-import app.dependency.update.app.model.dto.Plugins;
+import app.dependency.update.app.model.entities.Dependencies;
+import app.dependency.update.app.model.entities.NpmSkips;
+import app.dependency.update.app.model.entities.Packages;
+import app.dependency.update.app.model.entities.Plugins;
 import app.dependency.update.app.repository.DependenciesRepository;
 import app.dependency.update.app.repository.NpmSkipsRepository;
 import app.dependency.update.app.repository.PackagesRepository;
